pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "callguard"
version = "0.1.0"
description = "Runtime safety for AI agents. Stop agents before they break things."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

[project.optional-dependencies]
otel = ["opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20"]
langchain = ["langchain-core>=0.3"]
crewai = ["crewai>=0.80"]
agno = ["agno>=1.0"]
semantic-kernel = ["semantic-kernel>=1.0"]
openai-agents = ["openai-agents>=0.1"]
all = ["callguard[otel,langchain,crewai,agno,semantic-kernel,openai-agents]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "coverage>=7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/callguard"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
